+++ picasso/gui/filter.py
@@ -13,8 +13,6 @@ from __future__ import annotations
 
 import os.path
 import sys
-import importlib
-import pkgutil
 
 import numpy as np
 import pandas as pd
@@ -1072,20 +1070,11 @@ def main():
     app = QtWidgets.QApplication(sys.argv)
     window = Window()
 
-    from . import plugins
+    # load plugins from ~/.picasso/plugins
+    from .plugins_loader import load_plugins, add_plugins_menu_actions
 
-    def iter_namespace(pkg):
-        return pkgutil.iter_modules(pkg.__path__, pkg.__name__ + ".")
-
-    plugins = [
-        importlib.import_module(name)
-        for finder, name, ispkg in iter_namespace(plugins)
-    ]
-
-    for plugin in plugins:
-        p = plugin.Plugin(window)
-        if p.name == "filter":
-            p.execute()
+    load_plugins(window, "filter")
+    add_plugins_menu_actions(window, "filter")
 
     window.show()