+++ picasso/gui/render.py
@@ -16,8 +16,6 @@ import sys
 import copy
 import time
 import os.path
-import importlib
-import pkgutil
 from math import ceil
 from collections import Counter
 from functools import partial
@@ -13305,24 +13303,12 @@ class Window(QtWidgets.QMainWindow):
 def main():
     app = QtWidgets.QApplication(sys.argv)
     window = Window()
-    window.plugins = []
 
-    # load plugins from picasso/gui/plugins
-    from . import plugins
+    # load plugins from ~/.picasso/plugins
+    from .plugins_loader import load_plugins, add_plugins_menu_actions
 
-    def iter_namespace(pkg):
-        return pkgutil.iter_modules(pkg.__path__, pkg.__name__ + ".")
-
-    plugins = [
-        importlib.import_module(name)
-        for finder, name, ispkg in iter_namespace(plugins)
-    ]
-
-    for plugin in plugins:
-        p = plugin.Plugin(window)
-        if p.name == "render":
-            p.execute()
-            window.plugins.append(p)
+    load_plugins(window, "render")
+    add_plugins_menu_actions(window, "render")
 
     window.show()