+++ picasso/gui/simulate.py
@@ -15,8 +15,6 @@ import glob as _glob
 import os
 import sys
 import time
-import importlib
-import pkgutil
 import yaml
 from typing import Literal
 
@@ -2619,20 +2617,11 @@ def main():
     app = QtWidgets.QApplication(sys.argv)
     window = Window()
 
-    from . import plugins
+    # load plugins from ~/.picasso/plugins
+    from .plugins_loader import load_plugins, add_plugins_menu_actions
 
-    def iter_namespace(pkg):
-        return pkgutil.iter_modules(pkg.__path__, pkg.__name__ + ".")
-
-    plugins = [
-        importlib.import_module(name)
-        for finder, name, ispkg in iter_namespace(plugins)
-    ]
-
-    for plugin in plugins:
-        p = plugin.Plugin(window)
-        if p.name == "simulate":
-            p.execute()
+    load_plugins(window, "simulate")
+    add_plugins_menu_actions(window, "simulate")
 
     window.show()